+++ src/api/caidapp/urls.py
@@ -152,6 +152,7 @@ urlpatterns = [
         views.IdentityListView.as_view(),
         name="individual_identities",
     ),
+    path("dash_identities/", views.dash_identities, name="dash_identities"),
 
     path("init_identification/", views.init_identification, name="init_identification"),
     path("stop_init_identification/", views.stop_init_identification, name="stop_init_identification"),